jetson nano/requirements.txt
opencv-python>=4.5.5 
numpy>=1.19.5 
pillow>=8.4.0 
onnxruntime-gpu>=1.9.0 
torchvision>=0.8.2 
gtts>=2.2.3 
pygame>=2.1.0 
tensorrt>=8.2.0; 
python_version >= "3.6" 
# Optional, for TensorRT support pycuda>=2021.1; python_version >= "3.6" # Optional, for TensorRT support